Add files via upload
This project is an interactive analog clock created using HTML, CSS, and JavaScript. The clock design is visually appealing, with a classic analog style displayed on a webpage. The HTML structure defines the layout, while CSS is utilized for styling, including a background image of the clock face. JavaScript is employed to dynamically update the clock hands' positions based on the current time, ensuring accurate and real-time representation. The hands smoothly rotate to reflect the hours, minutes, and seconds, creating an engaging and functional user experience. The project not only serves as a practical application of web development skills but also showcases the integration of front-end technologies to create an aesthetically pleasing and responsive analog clock.